Disable staged install (fixes #1)
See:
* https://community.rstudio.com/t/no-staged-install-error/35321/2
* https://developer.r-project.org/Blog/public/2019/02/14/staged-install/index.html

This package creates variables to store filepaths that are calculated at
compilation time by `system.file()`.  That only works when the package isn't
moved anywhere after compilation, but the new default (staged install) is to
compile it in one place and then move it to another, to support parallel
compilation and installation.

It would be better to define the variables as functions, but I don't have time
now to check that it wouldn't break anything, so I'll disable staged
installations instead.